vmmon: avoid set_fs() in isVAReadable()
The set_fs() function is planned to be dropped from internal kernel API in
5.10 development cycle. In isVAReadable() it is used to check if a kernel
address can be read from without risking a pagefault if the answer is
negative by using set_fs() to pretend it's a userspace address and checking
with copy_from_user(). Starting with 5.8, copy_from_kernel_nofault() does
can be used to copy from a kernel address with access check but without an
actual pagefault.